removing more uses of dispatch
